Implemented price verification functionalty
Implemented price verification functionality that verifies the price sent from the client in the API request with the prices stored in the product database to protect against malicious users.   Also added the orderCreateReducer to the redux store.  Fixed the updateCartQuantity redux action to include the product description in its payload.