Home-Screen
Refactored the header navigation html to accomplish a more semantic and accessible layout. Removed the Favorite.jsx component as it was no longer being utilized. Refactored all the screens and components to include semicolons where necessary. Implemented functionality for the category links to filter the products on the product screen based on the category link clicked. Adjusted the ProductModel to include a category. Moved the Header component into the main app component instead of importing it to each screen.